Deleted `engine` and moved `Timer` & `UciLimit` to tools. Added better documentation.
