+++ pleco/src/bot_prelude.rs
@@ -7,4 +7,4 @@ pub use bots::AlphaBetaSearcher;
 pub use bots::JamboreeSearcher;
 pub use bots::IterativeSearcher;
 
-pub use engine::Searcher;
+pub use tools::Searcher;
