+++ pleco/src/bots/jamboree.rs
@@ -1,3 +1,4 @@
+//! The jamboree algorithm.
 use board::*;
 use core::piece_move::*;
 use board::eval::*;
