+++ pleco/src/bots/parallel_minimax.rs
@@ -1,3 +1,4 @@
+//! The parallel minimax algorithm.
 use board::*;
 use core::piece_move::*;
 use board::eval::*;
