+++ pleco/src/tools/timer.rs
@@ -1,7 +1,6 @@
 //! Timer for keeping track of the time for both sides of the a chess game.
 use std::time::Instant;
 
-// TODO: Clean this up.
 // Structure to keep track of time for two players.
 #[derive(Clone, Copy)]
 pub struct Timer {
@@ -18,10 +17,12 @@ enum Turn {
 }
 
 impl Timer {
+    /// Creates a new `Timer` with a set number of milliseconds left for each player.
     pub fn new_no_inc(milliseconds: i64) -> Self {
         Timer::new(milliseconds, milliseconds, 0, 0)
     }
 
+    /// Creates a new `Timer`.
     pub fn new(player_one_msec: i64, player_two_msec: i64, player_one_inc: i64, player_two_inc: i64) -> Self {
         Timer {
             start: Instant::now(),
@@ -31,11 +32,13 @@ impl Timer {
         }
     }
 
+    /// Returns the number of milliseconds remaining while running for the current turn.
     pub fn msec_remaining(&self) -> i64 {
         let diff = self.start.elapsed();
         self.milli_seconds_remaining[self.turn as usize] - diff.as_secs() as i64
     }
 
+    /// Returns the time increment for the current turn's player.
     pub fn current_time_inc(&self) -> i64 {
         match self.turn {
             Turn::One => self.inc[0],
@@ -43,6 +46,7 @@ impl Timer {
         }
     }
 
+    /// Returns the time increment for the current turn's opposing player.
     pub fn opponent_time_inc(&self) -> i64 {
         match self.turn {
             Turn::One => self.inc[1],
@@ -50,14 +54,17 @@ impl Timer {
         }
     }
 
+    /// Returns the number of milliseconds remaining for the opponent.
     pub fn opp_time_remaining(&self) -> i64 {
         self.milli_seconds_remaining[other_turn(self.turn) as usize]
     }
 
+    /// Starts the timer.
     pub fn start_time(&mut self) {
         self.start = Instant::now();
     }
 
+    /// Stops the timer.
     pub fn stop_time(&mut self) {
         let diff = self.start.elapsed();
         self.milli_seconds_remaining[self.turn as usize] -= diff.as_secs() as i64;
@@ -66,10 +73,12 @@ impl Timer {
         }
     }
 
+    /// Switches the turn.
     pub fn switch_turn(&mut self) {
         self.turn = other_turn(self.turn);
     }
 
+    /// Returns if the current player is out of time.
     pub fn out_of_time(&self) -> bool {
         self.milli_seconds_remaining[self.turn as usize] <= 0
     }
